+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-inventory
 
-## 2.0.0 (IN PROGRESS)
+## [2.0.0](https://github.com/folio-org/ui-inventory/tree/v2.0.0) (2020-03-17)
+[Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.1...v2.0.0)
 
 * Increase number of digits for HRID settings page. Refs UIIN-904
 * Add warning modal to HRID settings page. Refs UIIN-861